requirements.txt
asgiref==3.4.1
backcall==0.2.0
black==21.11b1
click==8.0.3
colorama==0.4.4
decorator==5.1.0
Django==4.0rc1
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
ipdb==0.13.9
ipython==7.30.1
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.24
Pygments==2.10.0
PyJWT==2.3.0
pytz==2021.3
regex==2021.11.10
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
traitlets==5.1.1
typing-extensions==4.0.0
tzdata==2021.5
wcwidth==0.2.5
